requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==3.0.0
async-lru==2.0.4
asyncio==3.4.3
attrs==24.3.0
bcrypt==4.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
cryptography==44.0.0
DateTime==5.5
DAWG-Python==0.7.2
debugpy==1.8.12
decorator==5.1.1
discord==2.3.2
discord.py==2.4.0
docopt==0.6.2
executing==2.2.0
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.2
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.5.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2
